Fix client_info KeyError due to static OP config.
In the CLIENTS config, one valid combination of keys is
client_registration and provider_info. This applies to OPs that support
neither discovery[1] nor dynamic client registration[2].

The issue is that instantiating OIDCClients with a config that includes
client_registration and provider_info keys (but not client_info) throws
a KeyError exception. This is because the code references
kwargs["client_info"], which does not exist.

This patch fixes the issue by referencing kwargs["client_registration"]
instead.

[1] https://openid.net/specs/openid-connect-discovery-1_0.html
[2] https://openid.net/specs/openid-connect-registration-1_0.html