Fix KeyError: registration_endpoint.
When working with an OP, such as Google, that has no registration
endpoint, rp3.py causes a KeyError to be thrown after successful
authentication with the OP.

On investigation this revealed a deeper issue, namely that rp3 was
looking up the client instance using the wrong key:

  File "./rp3.py", line 261, in application
    client = clients[session["op"]]

The key (session["op"]) is the issuer id of the OP, say
"https://accounts.google.com". However, the `clients` var is an instance
of OIDCClients, which stores its client instances using the key of the
associated configuration in CLIENTS, say, "google".

OIDCClients reacts to a lookup of a non-existing key by creating a
dynamic client and returning that instead. Since the OP (google) has no
registration endpoint, when the dynamic client tries to look it up, it
throws a KeyError on registration_endpoint.

This patch addresses the issue by using the `path` var as the key,
instead of `session["op"]` (issuer), resulting in the correct client
being found, as long as the `path` var is identical to the key to the
corresponding CLIENTS configuration; see the lines marked '***':

    CLIENTS = {
        "google": { # ***
            "srv_discovery_url": "https://accounts.google.com",
            "client_registration": {
                "client_id": "xxxxxxxx.apps.googleusercontent.com",
                "client_secret": "XXXXXXXXXXXXXXXXXXXXXXXX",
                "redirect_uris": ["http://localhost:8666/google"] # ***
            },
            # ...
        },
        # ...
    }

The `path` var is obtained from the trailing part of the redirect uri.

This may not be the intended mode of operation, but it does work.