Fix KeyError: scope.
This KeyError happens after successfully authenticating with the OP, and
the RP client wants to exchange the code for a token. One of the
parameters required for the call to the token endpoint is the scope,
which must match the scope in the original call to the authentication
endpoint.

The problem is that the code tries to get the scope from the
authentication response, but it is absent:

    'scope': response['scope']

This fixes the issue by getting the scope from the client
behavior:

    "scope": self.behaviour["scope"]